Updated the library command so that it uses the name in the destination target.
