--- .core/.cli/commands/reactium/component-lib/generator.js
@@ -1,26 +1,0 @@
-const ora = require('ora');
-const ActionSequence = require('action-sequence');
-
-module.exports = ({ action, params, props }) => {
-    const spinner = ora({
-        spinner: 'dots',
-        color: 'cyan',
-    });
-
-    spinner.start();
-
-    const actions = require('./actions')(spinner);
-
-    return ActionSequence({
-        actions,
-        options: { params, props },
-    })
-        .then(success => {
-            spinner.succeed('complete!');
-            return success;
-        })
-        .catch(error => {
-            spinner.fail('error!');
-            return error;
-        });
-};
