--- .core/.cli/commands/reactium/library/template/library.hbs
@@ -1,3 +1,0 @@
-import {{component}} from '{{from}}';
-
-export default {{component}};
