--- .core/manifest/templates/library.hbs
@@ -1,5 +1,0 @@
-// Generated by library-tools.js
-
-{{#each this}}
-export const {{this.domain}} = require('{{this.file}}').default;
-{{/each}}
