--- src/lib.js
@@ -1,3 +1,0 @@
-// Generated by library-tools.js
-
-export const Radio = require('components/common-ui/form/Radio/library').default;
